feat: enhance quiz and answer validation with new constraints and refactor length constants

+++ src/main/java/com/quizzo/dto/QuizAttemptDetailsResponse.java
@@ -1,7 +1,7 @@
 package com.quizzo.dto;
 
 public record QuizAttemptDetailsResponse(String name,
-                                         Float durationTime,
+                                         Integer durationTime,
                                          Integer questions,
                                          Integer eliminationsCount,
                                          Boolean multipleChoice) {
